requirements.txt
Flask==2.3.2
line-bot-sdk
SpeechRecognition
pydub
ffmpeg
google-cloud-speech
whisper
google-search-results